--- Assets/Resources/Prefabs/BattleUnits/UnitAction.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 534bb72d42e81994ea7c4c4c82e94f4f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
