--- Assets/Resources/Prefabs/BattleUnits/UnitAction/UnitAction.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7f241064d34c4bc4aa742a2261390c3e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
