--- Assets/Resources/Prefabs/BattleUnits/UnitAction/UnitAction_Iana.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 19b2c3c07c3ee8742ad883963d5a7d79
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
