--- Assets/Resources/Prefabs/UI/Scene/UI_BattleSpeedButton.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0bd8c66f81b96744f8541348da91c3b4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
