--- Assets/Resources/Prefabs/UI/Scene/UI_Hub.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3d496e3e51fa22646ab7729a923411f3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
