--- Assets/Script/Scenes/HubScene.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a7d9fdb37d775e24db8ba7492763e1f0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
