--- Assets/Script/Scenes/HubScene/HubSceneController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 54a547640571c8f429792bc6f93aea25
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
