--- Assets/Script/Scenes/MainScene/OpeningScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8e8824f87d8058248a4b916777a0e5bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
