--- Assets/Script/UI/Scene/UI_BattleSpeedButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ec7fe803bf408849b4a7bdf28a85d86
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
